nochmal...

+++ src/main/java/com/lennadi/eventbubble30/fileStorage/FileManagerService.java
@@ -1,24 +1,22 @@
 package com.lennadi.eventbubble30.fileStorage;
 
-import com.lennadi.eventbubble30.features.service.templates.ProfilePicture;
-import com.lennadi.eventbubble30.features.service.templates.StoredFile;
-import jakarta.validation.constraints.NotNull;
+import com.lennadi.eventbubble30.fileStorage.templates.StoredFile;
 import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 import org.springframework.web.multipart.MultipartFile;
 
 import java.io.IOException;
-import java.io.InputStream;
 import java.net.URL;
 import java.time.Instant;
 
+@Slf4j
 @Service
 @RequiredArgsConstructor
 public class FileManagerService {
     private final FileStorageService fileStorageService;
 
-    public ProfilePicture toProfilePicture(MultipartFile file) throws IOException {
-        ProfilePicture.validateUpload(file.getContentType(), file.getSize());
+    public StoredFile toProfilePicture(MultipartFile file) throws IOException {
 
         String newKey = fileStorageService.store(
                 file.getInputStream(),
@@ -26,9 +24,9 @@ public class FileManagerService {
                 file.getContentType()
         );
 
-        ProfilePicture ret = null;
+        StoredFile ret = null;
         try{
-            ret = ProfilePicture.ofStoredKey(newKey);
+            ret = StoredFile.ofStoredKey(newKey);
         }catch(Exception e){
             fileStorageService.delete(newKey);
         }
@@ -36,12 +34,16 @@ public class FileManagerService {
         return ret;
     }
 
-    public URL getURL(@NotNull StoredFile storedFile, int ttlSec) {
+    public URL getURL(StoredFile storedFile, int ttlSec) {
         Instant then =  Instant.now().plusSeconds(ttlSec);
         return getURL(storedFile, then);
     }
 
-    public URL getURL(@NotNull StoredFile storedFile, Instant expiry) {
+    public URL getURL(StoredFile storedFile, Instant expiry) {
+        if(storedFile == null){
+            log.warn("stored file is null");
+            return null;
+        }
         try{
             return fileStorageService.getFileURL(storedFile.getKey());//todo expiry
         }catch(Exception e){
@@ -50,11 +52,15 @@ public class FileManagerService {
         }
     }
 
-    public URL getURL(@NotNull StoredFile storedFile) {
+    public URL getURL(StoredFile storedFile) {
         return getURL(storedFile, null);
     }
 
-    public void delete(@NotNull StoredFile file) {
+    public void delete(StoredFile file) {
+        if(file == null){
+            log.warn("trying to delete null-file");
+            return;
+        }
         fileStorageService.delete(file.getKey());
     }
 }
